packages/training/pyproject.toml
[project]
name = "rhoai-mcp-training"
version = "0.1.0"
description = "Training component for RHOAI MCP - Kubeflow Training Operator integration"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Adam Miller", email = "admiller@redhat.com" }
]
maintainers = [
    { name = "Training Team", email = "training-team@redhat.com" }
]
requires-python = ">=3.10"
keywords = ["mcp", "openshift", "rhoai", "training", "kubeflow"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "rhoai-mcp-core>=0.1.0,<1.0.0",
]

[project.optional-dependencies]
kubeflow = [
    "kubeflow-training>=0.2.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.entry-points."rhoai_mcp.plugins"]
training = "rhoai_mcp_training.plugin:create_plugin"

[project.urls]
Homepage = "https://github.com/admiller/rhoai-mcp-prototype"
Repository = "https://github.com/admiller/rhoai-mcp-prototype"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rhoai_mcp_training"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint.isort]
known-first-party = ["rhoai_mcp_training", "rhoai_mcp_core"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
